Fix : Repository 인자 Application -> getPreference 로 바꾸기 데이터 관리와 테스트를 위해 명확하게 구분 지어놓음
